Move Couting Elements files to their package
