+++ src/main/java/algorithms/curated170/easy/countingelements/CountingElementsLinear.java
@@ -1,18 +1,14 @@
-package algorithms.curated170.easy;
+package algorithms.curated170.easy.coutingelements;
 
-import java.util.HashSet;
-import java.util.Set;
-
-public class CountingElementsSet {
+public class CountingElementsLinear {
     public int countElements(int[] arr) {
-
-        Set<Integer> nums = new HashSet<>();
+        boolean[] check = new boolean[1002];
         int count = 0;
         for (int n : arr) {
-            nums.add(n);
+            check[n] = true;
         }
         for (int n : arr) {
-            if (nums.contains(n + 1)) {
+            if (check[n + 1]) {
                 count++;
             }
         }
@@ -21,7 +17,7 @@ public class CountingElementsSet {
 
     public static void main(String[] args) {
 
-        var solution0 = new CountingElementsSet();
+        var solution0 = new CountingElementsLinear();
         var solution1 = new CountingElements();
 
         int[] arr = { 1, 2, 3 };
